+++ packages/e2e-tests/test-applications/node-express-app/src/app.ts
@@ -17,7 +17,7 @@ Sentry.init({
 });
 
 const app = express();
-const port = process.env.PORT ? parseInt(process.env.PORT) : 3000;
+const port = Number(process.env.BASE_PORT) + Number(process.env.PORT_MODULO);
 
 app.use(Sentry.Handlers.requestHandler());
 app.use(Sentry.Handlers.tracingHandler());