+++ packages/nextjs/src/server/utils/wrapperUtils.ts
@@ -1,7 +1,6 @@
-import { captureException, getActiveTransaction, getCurrentHub, startTransaction } from '@sentry/core';
+import { captureException, getActiveTransaction, runWithAsyncContext, startTransaction } from '@sentry/core';
 import type { Transaction } from '@sentry/types';
 import { baggageHeaderToDynamicSamplingContext, extractTraceparentData } from '@sentry/utils';
-import * as domain from 'domain';
 import type { IncomingMessage, ServerResponse } from 'http';
 
 import { platformSupportsStreaming } from './platformSupportsStreaming';
@@ -75,7 +74,7 @@ export function withTracedServerSideDataFetcher<F extends (...args: any[]) => Pr
   },
 ): (...params: Parameters<F>) => Promise<ReturnType<F>> {
   return async function (this: unknown, ...args: Parameters<F>): Promise<ReturnType<F>> {
-    return domain.create().bind(async () => {
+    return runWithAsyncContext(async hub => {
       let requestTransaction: Transaction | undefined = getTransactionFromRequest(req);
       let dataFetcherSpan;
 
@@ -134,7 +133,7 @@ export function withTracedServerSideDataFetcher<F extends (...args: any[]) => Pr
         });
       }
 
-      const currentScope = getCurrentHub().getScope();
+      const currentScope = hub.getScope();
       if (currentScope) {
         currentScope.setSpan(dataFetcherSpan);
         currentScope.setSDKProcessingMetadata({ request: req });
@@ -154,7 +153,7 @@ export function withTracedServerSideDataFetcher<F extends (...args: any[]) => Pr
           await flushQueue();
         }
       }
-    })();
+    });
   };
 }